Only show Show All button when on a subroute
